bugfix #1731 - ID3v2 comment descriptions hidden
Multiple ID3v2 comments may be present, with different descriptions.
iTunes uses this for storing much non-human-readable data (instead of
using TXX frames). TXX[X] and COM[M] frames are now returned with the
frame description as the array key, if present, or a numeric key if no
description is provided. Generally this means that human-readable
comments will have a numeric array key.