bugfix of a076a3ecee14ae5ca9b3b60dfdc0ba71c91bd7fe
The bugfix (G:22 ID3v2 TXXX description encoding) had a regression.
If you edit write a new tag in the file with merge_existing_data, the
description is not reencoded back to the original encoding, and the resulting
file is broken.

The issue with my fix is that you want to add a new TXXX with UTF-16 encoding
but use a default UTF-8 encoding, you have to encode description in UTF-8 and
the value in UTF-16

A better fix should be to change the original fix to alter only the decoded
frame, not the original ID3v2 one.