When parsing Quicktime subatoms, check the data length against the expected length.
Sometimes a container will contain slightly more data than expected, enough to read the header of an atom which is actually outside of the container. This guards against that. (The atom will be correctly read as parallel to the container in the next pass).

Example: a 'moov' atom with the first 8 bytes of a subsequent 'uuid' atom that appears afterward.