Fix MP3 CBR detection for non LAME "Info" frames.
The current code starts off by assuming that frames with "Xing" or
"Info" are VBR. For LAME files it then detects if it was really a CBR
file based on the VBR method. However, for non LAME files (such as
produced by ffmpeg/libavformat) which contain an "Info" frame they
are incorrectly detected as VBR even when they are CBR.

This patch adds an else block after the LAME block to force CBR for
"Info" frames. However, in some cases simply forcing CBR will trigger
a warning of corruptd data as the bitrate in the "Info" frame does
not necisarily match the bitrate in the rest of the actual audio
frames. So the new else case restarts the deccode by calling
decodeMPEGaudioHeader again with the offset adjusted by the size of
the "Info" frame.

I have not found any LAME CBR files where the bitrate in the "Info"
frame is different from the audio frames but I also updated the LAME
case to share the same logic as the non LAME case.