Remove redundant (and buggy) VBR bit rate calculation block.
The results of this block of code always seem to be over written by
similar code further down in the decodeMPEGaudioHeader function. Although
the second version of the code did not set the framelength so a framelength
calculation was added there.

The removed code also calculates the wrong bit rate for non stereo files
(the calculations should not vary depending on number of channels).