Fix an error when scanning WAV file via externally opened file pointer
The fixed problem is similar to the one reported earlier in
https://github.com/JamesHeinrich/getID3/issues/242. The problem was
fixed for MP3 and several other file types in 276fd3e5 but the WAV
files were missed at that time. Since then, also support for DSDIFF
files has been added with the same problem.

To recap, the problem case is when the file being analyzed is passed
to getID3 as a file pointer without giving a fully qualified file path.
When the file type specific module then spawns a secondary instance of
getID3, it has to pass it also the filesize given by the client
application. If null is passed as filesize to getID3::openfile, then it
tries to obtain the filesize from the system using the filename. But
this will fail if the filename is not fully qualified.